Replaced point Lists with LinkedLists
Indexing of lists was inefficient and the List class limited us to maximum 2 billion points. LinkedLists are unlimited in length and we can (hopefully) be more efficient in their use by passing nodes around, preventing the need to search through the list (in as much as this is possible).

+++ domainManager.cs
@@ -703,7 +703,7 @@ namespace LGTracer
             return (xVals, yVals, pressureVals, massSurplus);
         }
 
-        public (double[], double[], double[]) MapRandomToXYP( int nPoints, System.Random rng )
+        public (double[], double[], double[]) MapRandomToXYP( long nPoints, System.Random rng )
         {
             // Scatter randomly throughout domain
             double xSpan = XMax - XMin;
@@ -716,7 +716,7 @@ namespace LGTracer
             double[] xInitial = new double[nPoints];
             double[] yInitial = new double[nPoints];
             double[] pressureInitial = new double[nPoints];
-            for (int i=0; i<nPoints; i++)
+            for (long i=0; i<nPoints; i++)
             {
                 xInitial[i] = rng.NextDouble()*xSpan + xStart;
                 yInitial[i] = rng.NextDouble()*ySpan + yStart;
